sdcard/rootfs/root/pwnagotchi/scripts/requirements.txt
Crypto
requests
pyyaml
scapy
gym
stable-baselines
tweepy
file_read_backwards
numpy
